feat: place 이미지 지정하지 않았을 때 디폴트 이미지 선택, 유틸 함수로 분리 (#210)
